// client/src/app/activities/activity-interface.ts
export interface Activity {
    activityName: string;
    // Use an object if you need to store coordinates or similar structured data
    // location: {           
    //     lat: number;
    //     lng: number;
    // };
    activityType: string;
    surfaceType: string;
    difficultyLevel: "Beginner" | "Intermediate" | "Advanced";
    _id?: string;
}